Fix MathUtil conversion regressions on non-validation paths (#97)
* Fix MathUtil conversion regressions on non-validation paths

5.4.0 replaced several silently-truncating BigDecimal conversions with
MathUtil equivalents that throw. On the parse and validation paths that is
the intent, but four of the call sites are not validation paths:

- JavaDecimalHolder.value_hash_code() applied the max-number-chars limit,
  so hashCode() threw IllegalArgumentException for any decimal over the
  limit, poisoning every HashMap or HashSet holding the object. The limit
  bounds parse-time work and cannot be applied here anyway: the equal
  xsd:integer hashes without a limit, and the hashes have to agree. Guard
  the one case that is genuinely expensive instead - expanding a large
  negative scale, eg 1E+2000000000 - and hash such values from their
  canonical form. The branch is taken on integer-digit count, which is the
  same for every representation of a value, so equal values still agree.

- XmlObjectBase.getBigIntegerValue() and set_BigDecimal() in the int, long
  and integer holders let MathUtil's plain IllegalArgumentException escape,
  where the lexical path for the same value reports it as
  XmlValueOutOfRangeException. Translate it in one place so both paths and
  both magnitude regimes agree.

- GDurationBuilder.normalize() converted the whole-second carry with
  toInt() while GDateBuilder does the same computation with toLong() into
  the same long variable, so a fraction between int and long range threw
  in one and not the other.


* Bound the integral setters by their own range, not the parse limit

set_BigDecimal in the int and long holders applied
maxNumberOfCharsForNumbers, which bounds numbers read out of a document
and has no business gating a value handed to a setter directly: under
maxChars(8), setBigDecimalValue(123456789012) was rejected while
setBigIntegerValue and setLongValue accepted the same value.

The real constraint is the type's own range, which set_BigInteger already
enforces, so pass the width of that range - 19 digits for long, 10 for
int, at either end, as precision() ignores the sign. MathUtil still does
the conversion, so both expansion traps stay guarded: a large negative
scale is rejected before it is expanded, and a value below 1 truncates to
zero without computing the divisor. xs:integer is unbounded and keeps the
configured limit, which is its only bound.

Also drop stripTrailingZeros() from the oversized-value branch of
value_hash_code(): it is quadratic in the number of trailing zeros, so
hashing a 200000-digit value took 39 seconds. The digit count and sign
are just as representation-invariant and are cheap to read.


---------

+++ src/main/java/org/apache/xmlbeans/GDurationBuilder.java
@@ -350,7 +350,7 @@ public class GDurationBuilder implements GDurationSpecification, java.io.Seriali
         if (_fs != null && (_fs.signum() < 0 || _fs.compareTo(GDate._one) >= 0)) {
             BigDecimal bdcarry = _fs.setScale(0, RoundingMode.FLOOR);
             _fs = _fs.subtract(bdcarry);
-            carry = MathUtil.toInt(bdcarry);
+            carry = MathUtil.toLong(bdcarry);
         }
 
         if (carry != 0 || _s < 0 || _s > 59 || _m < 0 || _m > 50 || _h < 0 || _h > 23) {
